Refactored RecipeList and NavigationDrawer in CommonUi.kt. Added the MyRecipeScreen.kt
