+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
+    id 'kotlin-kapt'
 }
 
 android {
@@ -65,7 +66,7 @@ dependencies {
     debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_ui_version"
 
     //Staggered Grid
-    implementation 'androidx.compose.foundation:foundation:1.3.0'
+    //implementation 'androidx.compose.foundation:foundation:1.3.0'
 
     //navigation
     implementation "androidx.navigation:navigation-compose:2.5.3"
@@ -81,6 +82,7 @@ dependencies {
     // Room
     implementation "androidx.room:room-runtime:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
 
     // Coil
     implementation 'io.coil-kt:coil-compose:2.2.2'
