+++ app/src/main/java/com/example/dishpedia/data/MyRecipe.kt
@@ -1,18 +1,37 @@
 package com.example.dishpedia.data
 
-import android.media.Image
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import androidx.room.TypeConverter
+import androidx.room.TypeConverters
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+
 
 @Entity(tableName = "myRecipe")
+@TypeConverters(Converters::class)
 data class MyRecipe(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
-    val image: Image,
+    val image: String,
     val title: String = "",
     val summary: String = "",
     val readyInMinutes: Int = 0,
     val servings: Int = 0,
-    val ingredient: List<String> = listOf(""),
-    val instructions: List<String> = listOf("")
+    val ingredient: ArrayList<String>,
+    val instructions: ArrayList<String>
 )
+
+object Converters {
+    @TypeConverter
+    fun fromString(value: String?): ArrayList<String> {
+        val listType = object : TypeToken<ArrayList<String?>?>() {}.type
+        return Gson().fromJson(value, listType)
+    }
+
+    @TypeConverter
+    fun fromArrayList(list: ArrayList<String?>?): String {
+        val gson = Gson()
+        return gson.toJson(list)
+    }
+}
