+++ app/src/main/java/com/example/dishpedia/models/NavigationDrawerItems.kt
@@ -4,3 +4,21 @@ class NavigationDrawerItems (
     val name: String,
     val route: String,
 )
+
+object NavigationDrawerItemsProvider{
+    val navItems = listOf(
+        NavigationDrawerItems(
+            name = "Home",
+            route = "Home",
+        ),
+        NavigationDrawerItems(
+            name = "My Recipes",
+            route = "MyRecipes",
+        ),
+        //TODO: Add authorization, then accounts
+//        NavigationDrawerItems(
+//            name = "Account",
+//            route = "Account",
+//        )
+    )
+}
