+++ app/src/main/java/com/example/dishpedia/ui/screens/SearchScreen.kt
@@ -1,33 +1,35 @@
 package com.example.dishpedia.ui.screens
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.*
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.OutlinedTextField
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.material.TextFieldDefaults
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Clear
 import androidx.compose.material.icons.rounded.Search
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
-import com.example.dishpedia.DishpediaScreen
-import com.example.dishpedia.R
-import com.example.dishpedia.models.Recipe
-import com.example.dishpedia.models.Recipes
+import com.example.dishpedia.utils.RecipeList
 import com.example.dishpedia.viewmodel.RecipesUiState
 import com.example.dishpedia.viewmodel.RecipesViewModel
 
@@ -99,58 +101,6 @@ fun SearchBar(recipesViewModel: RecipesViewModel){
             unfocusedBorderColor = Color.Transparent
         )
     )
-}
-
-@Composable
-private fun RecipeCard(
-    recipe: Recipe,
-    recipesViewModel: RecipesViewModel,
-    navController: NavController,
-    modifier: Modifier = Modifier
-){
-    Card(
-        modifier = modifier
-            .padding(8.dp)
-            .fillMaxWidth()
-            .clickable {
-                recipesViewModel.getRecipeById(recipe.id)
-                navController.navigate(DishpediaScreen.RecipeInfo.name)
-            }
-        ,
-        elevation = 4.dp
-    ) {
-        Column {
-            AsyncImage(
-                model = ImageRequest.Builder(context = LocalContext.current)
-                    .data(recipe.image)
-                    .crossfade(true)
-                    .build(),
-                contentDescription = recipe.title,
-                modifier = Modifier.fillMaxWidth(),
-                error = painterResource(id = R.drawable.ic_connection_error),
-                placeholder = painterResource(id = R.drawable.loading_img),
-                contentScale = ContentScale.FillWidth
-            )
-            Text(
-                text = recipe.title,
-                modifier = Modifier.padding(16.dp),
-                style = MaterialTheme.typography.h6
-            )
-        }
-    }
-}
-
-@Composable
-private fun RecipeList(
-    recipes: Recipes,
-    recipesViewModel: RecipesViewModel,
-    navController: NavController
-){
-    LazyColumn{
-        items(recipes.recipes){recipe ->
-            RecipeCard(recipe, recipesViewModel, navController)
-        }
-    }
 }
 
 //@Preview(showBackground = true)
